# src/api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Agent API",
    "description": "ForU AI Agent API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/personality-to-character": {
      "post": {
        "summary": "Generate a character profile from personality data",
        "description": "Processes given agent data and returns a formatted character profile.",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key required for authentication."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agent_name": {
                    "type": "string",
                    "example": "Eliza"
                  },
                  "agent_personality": {
                    "type": "string",
                    "example": "Friendly and helpful"
                  },
                  "traits": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["curious", "analytical", "empathetic"]
                  },
                  "tone_voice": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["formal", "professional"]
                  },
                  "community_name": {
                    "type": "string",
                    "example": "Tech Enthusiasts"
                  },
                  "community_description": {
                    "type": "string",
                    "example": "A community for discussing the latest technology trends."
                  }
                },
                "required": [
                  "agent_name",
                  "agent_personality",
                  "traits",
                  "tone_voice",
                  "community_name",
                  "community_description"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request. Invalid input data."
          },
          "401": {
            "description": "Unauthorized. Missing or invalid API key."
          },
          "500": {
            "description": "Internal server error. Something went wrong while processing the request."
          }
        }
      }
    },
    "/agents": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/agents/{agentId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/agents/{agentId}/set": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agents-create": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agentId}/{roomId}/memories": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{agentId}/message": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roomId": {
                    "example": "any"
                  },
                  "userId": {
                    "example": "any"
                  },
                  "userName": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "text": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{agentId}/image": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}